feat: Update dependencies and configurations
This commit updates the dependencies and configurations of the project. It adds the `image` package with version 4.2.0 to the `pubspec.yaml` file, updates the iOS platform version to 14.0 in the `Podfile`, and adds the `NSLocationWhenInUseUsageDescription` key to the `AppFrameworkInfo.plist` file. These changes improve the functionality and compatibility of the app with the latest libraries and platforms.