chore: Update Java and Gradle versions
This commit updates the Java and Gradle versions in the build.gradle and gradle-wrapper.properties files. The sourceCompatibility, targetCompatibility, and jvmTarget have been changed to JavaVersion.VERSION_17 to support the latest Java features. The kotlin_version has been updated to '2.0.10' and the gradle_version has been updated to '8.3.0' to ensure compatibility with the latest Kotlin and Gradle versions. These updates improve the development environment and enable the use of newer language features and tools.