chore: Update Android build.gradle files
This commit updates the Android build.gradle files to fix an error related to the namespace. The changes include updating the ndkVersion to "25.1.8937393" in the app/build.gradle file and adding code to fix the error in the allprojects section of the build.gradle file. These updates ensure the correct configuration and improve the build process for the Android app.