chore: Update global_var.dart and dashboard.dart
This commit updates the global_var.dart file to add a new variable `timerDuration` with a value of 20. This variable will be used for the duration of a timer in the NotificationDialog widget.

In the dashboard.dart file, the code has been updated to handle errors when refreshing the user. If the app is in debug mode, the error message will be printed to the console. This change improves error handling and debugging capabilities.