feat: Add isDriverAvailable and isDriverAvailableServerSide variables
This commit adds the `isDriverAvailable` and `isDriverAvailableServerSide` variables to the `global_var.dart` file. These variables are used to track the availability status of the driver. The `isDriverAvailable` variable is used client-side, while the `isDriverAvailableServerSide` variable is used server-side. This change enhances the functionality of the app by providing a more accurate representation of the driver's availability.