refactor: Update FirestoreMethods class to include getTrips method
This commit adds the getTrips method to the FirestoreMethods class in firestore_methods.dart. The method retrieves the trips of the user from the earnings collection in Firestore and returns a list of EndedTripDetails objects. This change enhances the functionality of the app by allowing users to access their trip history and details.