refactor: Update profile screen UI and signout functionality
This commit updates the `_ProfileScreenState` class in `profile_screen.dart` to improve the UI of the profile screen. It adds rounded borders to the cards displayed on the screen, enhancing the visual appeal and providing a consistent design across the app. Additionally, it updates the signout functionality by calling the `goOfflinePermanently` method and then signing out the user. These changes enhance the user experience and ensure a smoother transition when signing out.